Data types for in-kernel rxgk tokens
We will only be able to use the GetTokens2 and SetTokens2 variants,
which use XDR to encode/decode data across the pioctl boundary.
Since userspace only needs to use these types immediatly around pioctl,
it can use the XDR-produced types directly, but these data types are
rather clunky for regular use in the kernel.

As such, define additional types for use in the kernel, as is done
for the existing rxkad functionality.

The XDR types for pioctl use should only be used for dbserver
tokens; userland should not be acquiring fileserver tokens for
kernel use, and the kernel should not be caching fileserver tokens
where they might get returned to userland.  This allows us to
preserve the invariant that a TokenJar has only one token per
user per cell per token type.

In order to make the XDR types for rxgk tokens dependent on the
configure option --enable-rxgk, we must include afsconfig.h in
token.xg for use by rxgen's preprocessor.  This requires some
tweaks to work everywhere, since afsconfig.h can bring in sys/types.h,
which on some systems has typedefs that are not compatible with
RPC-L typedefs.