Add gklog; insert rxgk tokens into the kernel
This is mostly a stub implementation with very minimal
functionality.  All of the nice things that aklog provides
(log to paths, log to multiple cells, .xlog, etc.) are not
supported; this is just a way to do functionality testing
at the moment.

Keep the old copyright from the aklog code this started from, and
update the date for the new work.  (Same rightsholder, imagine that.)

Ideally it would be possible to factor out most of aklog into
something that can be reused for rxgk, but that is probably a
fair amount of work.

Change-Id: I83c9a6b5646dcbaa692df535ba3965b8b4309f2d