Make libtool export lists generated files
Rename *.la.sym to .la.sym.in so that configure can generate the
actual .la.sym files to be used.  Adjust the libtool library build
rules accordingly to find the export symbol list in the build
directory.

This is needed because some rxgk functionality provided by libafsrpc
will not be available when configured without rxgk.

Change-Id: I07be3a732124974fc3eb26fc30362aa963d9cdd4