Remove incomplete struct initializers for rxkad stats
C99 requires that objects with static linkage, which includes
global variables, be initialized to zero/NULL.

It is possible that old compilers required a hack of using one
explicit initializer and relying on the requirement from C99 that
the elements of the structure not listed in the initializer be
initialized as if it had static linkage.

Using a complete explicit initializer is undesired here, as the
statistics structures have a great number of elements.

Change-Id: Iefe7842cbf874252267cb3a8aee5d90ec2cab169