Build hcrypto with libtool
XXX changes LIB_roken and LIB_hcrypto in hacky ways.

Or rather, with lwptool, since we need a LWP version as well as
the various pthreaded versions.

The previous version was the initial version, 1.1, but since we're
switching to libtool, bump the version to 2.0 just to be safe.

Libtool abstracts away the extra logic that had previously been needed
to build different copies of rand-fortuna for the pthreaded and LWP
libraries.

Change-Id: Ibc530a1fa4baa7a38b44eb3e0719e1905a6fe269