Build libuafs with lwptool
Use the standard program for building slightly different object files
into their own directories, instead of rolling our own.  This allows
us to pull the build rules into the Makefile.common, leaving just
compiler flags and similar in the MakefileProtos.

We do have to hack around the standard libtool build rules a little
bit in order to get things compiled using the same compiler and arguments
that the old build system used, but it does not seem too bad.

As a side effect, we get rid of the LIBUAFS make variable -- all sites
were defining it to libuafs.a anyway, so we can just hardcode it.

Change-Id: I2bda2f40bbd0aa808c24e074d2d7bcd329f6b77e