Provide a usable fileserver default config (-X)
The existing default values, and the -S (small) and -L (large) presets,
are essentially untouched since the initial import from IBM, and as
such are scaled appropriately for 1995-era hardware.  (The number of
threads used by a large fileserver increased to 128 in 2008.)

Modern systems are several iterations of Moore's law ahead of those,
and can generally afford to use a lot more memory for data structures.
We should provide a usable preset configuration instead of requiring
all operators to make extensive tuning to their fileservers to get
decent performance.

Add the -X preset (extra large) for this purpose, since we do not
have sufficient internal sanity checking to prevent changing the
existing default/presets from having a harmful effect on existing
deployments which add extra customizations on top of a preset.

Update the man pages to add the -X preset and document that -S and
-L are deprecated.

Change-Id: I35751bab3acc4a9a72ec0d04e88f36d2c2c44298