Attempt to make the server install bits current
Avoid using -noauth, and mention both the rxkad.keytab (1.6)
and the KeyFileExt (as 1.8, though it's only master at present).
To support these, move forward the use of kadmin to extract
the afs/cell principal's keytab.

Move the buserver's creation to the end of the list and mark it
as optional (many sites do not run the AFS backup suite).

Deindent some programlisting blocks so they don't flow off the
page as much in the PDF version.

Drop vos syncserv and vos syncvldb from the tasks for setting
up a new server; they should not be needed, as the new db server
should pick up the existing database when it joins the quorum.

General edits for clarity, whitespace and such.

Change-Id: I951ec3ee237e4c83a17c82802328f0a454b61097