Try to make compile_et standalone
That is, not reliant on an OpenAFS build tree for things like
afsconfig.h.

The inclusion of afsconfig.h was added in commit
972a4072827fb2ec680354d5adebc2c5cca06939 to ensure that it was included
prior to afs/param.h; however, the inclusion of afs/param.h in
compile_et.c stems from the original IBM import and seems of minimal
value.  The only change needed to build without param.h is to use
int instead of afs_int32 in a couple places -- int is 32 bits on
all platforms currently supported.

Change-Id: I1ee969eec92b139d265a7494e13ddfc69c05f238