Linux 4.5: get_link instead of follow_link+put_link
In linux commit 6b255391, the follow_link inode operation was
replaced by the get_link operation, which is basically the same
but takes the inode and dentry separately, allowing for the
possibility of staying in RCU mode.

In linux commit fceef393a, the put_link() inode operation was removed,
since the necessary destructor can be set as a delayed call in the
get_link() operation.

Sometimes we can use a builtin get_link implementation, but not
when !defined(USABLE_KERNEL_PAGE_SYMLINK_CACHE), so we have to
adapt our callers accordingly.

Since configure tests for the absence of a field can suffer from
false positives and these changes were grouped together in
a common patch series (merge commit 32fb378), prefer checking
for get_link instead of put_link, as appropriate.