Linux: lock the parent dentry for lookup_one_len()
During review of the changes in Linux 4.5, it was noted that
callers of lookup_one_len() are obligated to hold the i_mutex
on the parent directory, though we do not currently do so.
Take the lock around the call as needed.

The only call to lookup_one_len() occurs in afs_linux_sillyrename(),
so may workloads may not encounter this codepath.